Expose add_case_closure
