+++ experimental/casemapping/src/casemapping.rs
@@ -5,6 +5,7 @@
 use crate::internals::{CaseMapLocale, FoldOptions};
 use crate::provider::data::MappingKind;
 use crate::provider::CaseMappingV1Marker;
+use crate::set::ClosureSet;
 use alloc::string::String;
 use icu_locid::Locale;
 use icu_provider::prelude::*;
@@ -250,5 +251,35 @@ impl CaseMapping {
             .full_helper_writeable(src, CaseMapLocale::Turkish, MappingKind::Fold)
             .write_to_string()
             .into_owned()
+    }
+
+    /// Adds all simple case mappings and the full case folding for `c` to `set`.
+    /// Also adds special case closure mappings.
+    ///
+    /// In other words, this adds all strings/characters that this casemaps to, as
+    /// well as all characters that may casemap to this one.
+    ///
+    /// The character itself is not added.
+    ///
+    /// For example, the mappings
+    /// - for s include long s
+    /// - for sharp s include ss
+    /// - for k include the Kelvin sign
+    pub fn add_case_closure<S: ClosureSet>(&self, c: char, set: &mut S) {
+        self.data.get().add_case_closure(c, set);
+    }
+
+    /// Maps the string to single code points and adds the associated case closure
+    /// mappings, if they exist.
+    ///
+    /// The string is mapped to code points if it is their full case folding string.
+    /// In other words, this performs a reverse full case folding and then
+    /// adds the case closure items of the resulting code points.
+    /// If the string is found and its closure applied, then
+    /// the string itself is added as well as part of its code points' closure.
+    ///
+    /// Returns true if the string was found
+    pub fn add_string_case_closure<S: ClosureSet>(&self, s: &str, set: &mut S) -> bool {
+        self.data.get().add_string_case_closure(s, set)
     }
 }
